examples: refactor the tensorflow-mnist-feature-squeezing example
Refactors the tensorflow-mnist-feature-squeezing demo to use the new, common approach designed for
the examples folder. This includes the following changes:

- Moved any new custom task plugins from tensorflow-mnist-feature-squeezing folder to the root of
  the examples folder.
- Moved the MLproject file into the src folder.
- Refactored the task plugins and entrypoint scripts to remove all direct usages of Prefect tasks.
- Pruned ununsed task plugins and entrypoint scripts from the example.
- Updated the documentation in examples/tensorflow-mnist-feature-squeezing/README.md with
  instructions on how the user should prepare their environment for running the examples.
- Modified the demo Jupyter notebook for tensorflow-mnist-feature-squeezing to use the new scripts.
- Removed any files from the tensorflow-mnist-feature-squeezing folder that were no longer being
  used after implementing the above changes.
