--- examples/tensorflow-mnist-feature-squeezing/.gitignore
@@ -1,2 +1,0 @@
-data/**/
-!data/download_data.py
