+++ examples/tensorflow-mnist-feature-squeezing/README.md
@@ -1,31 +1,22 @@
-# Tensorflow MNIST Feature Squeezing Demo
+# Tensorflow MNIST Feature Squeezing demo
 
-> ⚠️ **IMPORTANT!**
->
-> This README is out of date and will be updated in the near future.
->
-> There is a new setup tool that all users should use to configure and run Dioptra, please see the new sections [Building the containers](https://pages.nist.gov/dioptra/getting-started/building-the-containers.html) and [Running Dioptra](https://pages.nist.gov/dioptra/getting-started/running-dioptra.html) that have been added to the documentation.
+This example demonstrates using Dioptra to run experiments that investigate the effectiveness of the feature-squeezing defense against a series of evasion attacks against a neural network model.
+The demo can be found in the Jupyter notebook file [demo.ipynb](demo.ipynb).
 
->⚠️ **Warning:** Some of the attacks in this demo, _deepfool_ and _CW_ in particular, are computationally expensive and will take a very long to complete if run using the CPUs found in a typical personal computer.
-> For this reason, it is highly recommended that you run these demos on a CUDA-compatible GPU.
+## Running the example
 
-The demo provided in the Jupyter notebook `demo.ipynb` uses Dioptra to run experiments that investigate the effectiveness of the feature-squeezing defense against a series of evasion attacks against a neural network model.
+To prepare your environment for running this example, follow the linked instructions below:
 
-## Getting started
+1.  [Create and activate a Python virtual environment and install the necessary dependencies](../README.md#creating-a-virtual-environment)
+2.  [Download the MNIST dataset using the download_data.py script.](../README.md#downloading-datasets)
+3.  [Follow the links in these User Setup instructions](../../README.md#user-setup) to do the following:
+    -   Build the containers
+    -   Use the cookiecutter template to generate the scripts, configuration files, and Docker Compose files you will need to run Dioptra
+4.  [Edit the docker-compose.yml file to mount the data folder in the worker containers](../README.md#mounting-the-data-folder-in-the-worker-containers)
+5.  [Initialize and start Dioptra](https://pages.nist.gov/dioptra/getting-started/running-dioptra.html#initializing-the-deployment)
+6.  [Register the custom task plugins for Dioptra's examples and demos](../README.md#registering-custom-task-plugins)
+7.  [Register the queues for Dioptra's examples and demos](../README.md#registering-queues)
+8.  [Start JupyterLab and open `demo.ipynb`](../README.md#starting-jupyter-lab)
 
-Everything you need to run this demo is packaged into a set of Docker images that you can obtain by opening a terminal, navigating to the root directory of the repository, and running `make pull-latest`.
-Once you have downloaded the images, navigate to this example's directory using the terminal and run the demo startup sequence:
-
-```bash
-make demo
-```
-
-The startup sequence will take more time to finish the first time you use this demo, as you will need to download the MNIST dataset, initialize the Testbed API database, and synchronize the task plugins to the S3 storage.
-Once the startup process completes, open up your web browser and enter `http://localhost:38888` in the address bar to access the Jupyter Lab interface (if nothing shows up, wait 10-15 more seconds and try again).
-Double click the `work` folder and open the `demo.ipynb` file.
-From here, follow the provided instructions to run the demo provided in the Jupyter notebook.
-
-If you are running the demos locally and want to watch the output logs for the Tensorflow worker containers as you step through the demo, run `docker-compose logs -f tfcpu-01 tfcpu-02` in your terminal.
-
-When you are done running the demo, close the browser tab containing this Jupyter notebook and shut down the services by running `make teardown` on the command-line.
-If you were watching the output logs, you will need to press <kbd>Ctrl</kbd>-<kbd>C</kbd> to stop following the logs before you can run `make teardown`.
+Steps 1–4 and 6–7 only need to be run once.
+**Returning users only need to repeat Steps 5 (if you stopped Dioptra using `docker compose down`) and 8 (if you stopped the `jupyter lab` process)**.
