+++ examples/tensorflow-mnist-feature-squeezing/src/cw_inf.py
@@ -39,6 +39,7 @@ from dioptra.sdk.utilities.logging import (
     set_logging_level,
 )
 
+_CUSTOM_PLUGINS_IMPORT_PATH: str = "dioptra_custom"
 _PLUGINS_IMPORT_PATH: str = "dioptra_builtins"
 DISTANCE_METRICS: List[Dict[str, str]] = [
     {"name": "l_infinity_norm", "func": "l_inf_norm"},
@@ -135,18 +136,6 @@ def _coerce_int_to_bool(ctx, param, value):
     help=" Confidence of adversarial examples",
 )
 @click.option(
-    "--max-halving",
-    type=click.INT,
-    help="Maximum number of halving steps in the line search optimization",
-    default=5,
-)
-@click.option(
-    "--max-doubling",
-    type=click.INT,
-    help="Maximum number of doubling steps in the line search optimization",
-    default=5,
-)
-@click.option(
     "--seed",
     type=click.INT,
     help="Set the entry point rng seed",
@@ -178,8 +167,6 @@ def cw_inf_attack(
     targeted,
     learning_rate,
     max_iter,
-    max_doubling,
-    max_halving,
     verbose,
 ):
     LOGGER.info(
@@ -198,8 +185,6 @@ def cw_inf_attack(
         targeted=targeted,
         learning_rate=learning_rate,
         max_iter=max_iter,
-        max_doubling=max_doubling,
-        max_halving=max_halving,
         verbose=verbose,
     )
 
@@ -221,8 +206,6 @@ def cw_inf_attack(
                 targeted=targeted,
                 learning_rate=learning_rate,
                 max_iter=max_iter,
-                max_doubling=max_doubling,
-                max_halving=max_halving,
                 verbose=verbose,
             )
         )
@@ -247,8 +230,6 @@ def init_cw_flow() -> Flow:
             targeted,
             learning_rate,
             max_iter,
-            max_doubling,
-            max_halving,
             verbose,
         ) = (
             Parameter("testing_dir"),
@@ -265,8 +246,6 @@ def init_cw_flow() -> Flow:
             Parameter("targeted"),
             Parameter("learning_rate"),
             Parameter("max_iter"),
-            Parameter("max_doubling"),
-            Parameter("max_halving"),
             Parameter("verbose"),
         )
         seed, rng = pyplugs.call_task(
@@ -316,7 +295,7 @@ def init_cw_flow() -> Flow:
             request=DISTANCE_METRICS,
         )
         distance_metrics = pyplugs.call_task(
-            "src",
+            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.feature_squeezing",
             "cw_inf_plugin",
             "create_adversarial_cw_inf_dataset",
             model_name=model_name,
@@ -327,13 +306,11 @@ def init_cw_flow() -> Flow:
             adv_data_dir=adv_data_dir,
             batch_size=batch_size,
             image_size=image_size,
-            max_doubling=max_doubling,
             max_iter=max_iter,
             targeted=targeted,
             confidence=confidence,
             learning_rate=learning_rate,
             model_architecture=model_architecture,
-            max_halving=max_halving,
             verbose=verbose,
             initial_const=1,
             upstream_tasks=[make_directories_results],
