--- examples/tensorflow-mnist-feature-squeezing/src/data_res.py
@@ -1,81 +1,0 @@
-# This Software (Dioptra) is being made available as a public service by the
-# National Institute of Standards and Technology (NIST), an Agency of the United
-# States Department of Commerce. This software was developed in part by employees of
-# NIST and in part by NIST contractors. Copyright in portions of this software that
-# were developed by NIST contractors has been licensed or assigned to NIST. Pursuant
-# to Title 17 United States Code Section 105, works of NIST employees are not
-# subject to copyright protection in the United States. However, NIST may hold
-# international copyright in software created by its employees and domestic
-# copyright (or licensing rights) in portions of software that were assigned or
-# licensed to NIST. To the extent that NIST holds copyright in this software, it is
-# being made available under the Creative Commons Attribution 4.0 International
-# license (CC BY 4.0). The disclaimers of the CC BY 4.0 license apply to all parts
-# of the software developed or licensed by NIST.
-#
-# ACCESS THE FULL CC BY 4.0 LICENSE HERE:
-# https://creativecommons.org/licenses/by/4.0/legalcode
-import warnings
-from typing import Optional, Tuple
-
-warnings.filterwarnings("ignore")
-
-import tensorflow as tf
-
-from tensorflow.keras.applications.resnet50 import preprocess_input
-
-tf.compat.v1.disable_eager_execution()
-
-import structlog
-from mlflow.tracking import MlflowClient
-from tensorflow.keras.preprocessing.image import ImageDataGenerator
-
-
-LOGGER = structlog.get_logger()
-
-
-def download_image_archive(
-    run_id: str, archive_path: str, destination_path: Optional[str] = None
-) -> str:
-    client: MlflowClient = MlflowClient()
-    image_archive_path: str = client.download_artifacts(
-        run_id=run_id, path=archive_path, dst_path=destination_path
-    )
-    LOGGER.info(
-        "Image archive downloaded",
-        run_id=run_id,
-        storage_path=archive_path,
-        dst_path=image_archive_path,
-    )
-    return image_archive_path
-
-
-def create_image_dataset(
-    data_dir: str,
-    subset: str,
-    rescale: float = 1.0,
-    validation_split: float = 0.2,
-    batch_size: int = 32,
-    seed: int = 8237131,
-    label_mode: str = "categorical",
-    color_mode: str = "rgb",
-    image_size: Tuple[int, int] = (224, 224),
-    model_architecture: str = "resnet50",
-):
-    if model_architecture == "mobilenet":
-        from tensorflow.keras.applications.mobilenet import preprocess_input
-
-    data_generator: ImageDataGenerator = ImageDataGenerator(
-        rescale=rescale,
-        validation_split=validation_split,
-        preprocessing_function=preprocess_input,
-    )
-
-    return data_generator.flow_from_directory(
-        directory=data_dir,
-        target_size=image_size,
-        color_mode=color_mode,
-        class_mode=label_mode,
-        batch_size=batch_size,
-        seed=seed,
-        subset=subset,
-    )
